+++ starlight/src/route.rs
@@ -6,7 +6,6 @@ mod config;
 mod debug;
 mod embed;
 mod path;
-mod region_adv;
 mod router;
 mod routing;
 
@@ -16,6 +15,5 @@ pub use cnode::CNode;
 pub use config::{Config, Configurator, PConfig};
 pub use embed::{Embedding, EmbeddingKind, PEmbedding};
 pub use path::{Edge, EdgeKind, HyperPath, Path};
-pub use region_adv::RegionAdvancer;
 pub use router::{PCEdge, PCNode, QCEdge, QCNode, Router};
 pub(crate) use routing::route;